TUI branch

+++ src/nanorc/__main_timing__.py
@@ -27,6 +27,7 @@ from nanorc.core import NanoRC
 from nanorc.runmgr import DBRunNumberManager
 from nanorc.cfgsvr import DBConfigSaver
 from nanorc.credmgr import credentials
+from nanorc.rest import RestApi, NanoWebContext, rc_context
 from . import confdata
 import nanorc.argval as argval
 
@@ -45,12 +46,13 @@ from nanorc.nano_context import NanoContext
 @click.option('--kerberos/--no-kerberos', default=True, help='Whether you want to use kerberos for communicating between processes')
 @click.option('--partition-number', type=int, default=0, help='Which partition number to run', callback=argval.validate_partition_number)
 @click.option('--web/--no-web', is_flag=True, default=False, help='whether to spawn webui')
+@click.option('--tui/--no-tui', is_flag=True, default=False, help='whether to use TUI')
 @click.option('--pm', type=str, default="ssh://", help='Process manager, can be: ssh://, kind://, or k8s://np04-srv-015:31000, for example', callback=argval.validate_pm)
 @click.argument('cfg_dir', type=str, callback=argval.validate_conf)
 @click.argument('partition-label', type=str, callback=argval.validate_partition)
 @click.pass_obj
 @click.pass_context
-def timingcli(ctx, obj, traceback, pm, loglevel, log_path, cfg_dumpdir, kerberos, timeout, partition_number, partition_label, web, cfg_dir):
+def timingcli(ctx, obj, traceback, pm, loglevel, log_path, cfg_dumpdir, kerberos, timeout, partition_number, partition_label, web, tui, cfg_dir):
     obj.print_traceback = traceback
     credentials.user = 'user'
     ctx.command.shell.prompt = f"{credentials.user}@timingrc> "
@@ -61,7 +63,8 @@ def timingcli(ctx, obj, traceback, pm, loglevel, log_path, cfg_dumpdir, kerberos
     grid.add_row("  but trust it and it will betray you!")
     grid.add_row(f"Use it with care, {credentials.user}!")
 
-    obj.console.print(Panel.fit(grid))
+    if not tui:
+        obj.console.print(Panel.fit(grid))
 
     port_offset = 300 + partition_number * 500
     rest_port = 5005 + partition_number
@@ -72,6 +75,8 @@ def timingcli(ctx, obj, traceback, pm, loglevel, log_path, cfg_dumpdir, kerberos
 
     rest_thread  = threading.Thread()
     webui_thread = threading.Thread()
+    if web and tui:
+        raise RuntimeError("cant have TUI and GUI at the same time")
     try:
         rc = NanoRC(
             console = obj.console,
@@ -89,7 +94,7 @@ def timingcli(ctx, obj, traceback, pm, loglevel, log_path, cfg_dumpdir, kerberos
 
         rc.log_path = os.path.abspath(log_path)
         add_custom_cmds(ctx.command, rc.execute_custom_command, rc.custom_cmd)
-        if web:
+        if web or tui:
             host = socket.gethostname()
 
             rc_context.obj = obj
@@ -105,30 +110,32 @@ def timingcli(ctx, obj, traceback, pm, loglevel, log_path, cfg_dumpdir, kerberos
             rest_thread.start()
             obj.console.log(f"Started RESTAPI")
 
-            webui_thread = None
-            obj.console.log(f'Starting up Web UI on {host}:{webui_port}')
-            webui = WebServer(host, webui_port, host, rest_port)
-            webui_thread = threading.Thread(target=webui.run, name='NanoRC_WebUI')
-            webui_thread.start()
-            obj.console.log(f"")
-            obj.console.log(f"")
-            obj.console.log(f"")
-            obj.console.log(f"")
-            grid = Table(title='Web NanoRC', show_header=False, show_edge=False)
-            grid.add_column()
-            grid.add_row(f"Started Web UI, you can now connect to: [blue]{host}:{webui_port}[/blue],")
-            if 'np04' in host:
-                grid.add_row(f"You probably need to set up a SOCKS proxy to lxplus:")
-                grid.add_row("[blue]ssh -N -D 8080 your_cern_uname@lxtunnel.cern.ch[/blue] # on a different terminal window on your machine")
-                grid.add_row(f'Make sure you set up browser SOCKS proxy with port 8080 too,')
-                grid.add_row('on Chrome, \'Hotplate localhost SOCKS proxy setup\' works well).')
-            grid.add_row()
-            grid.add_row(f'[red]To stop this, ctrl-c [/red][bold red]twice[/bold red] (that will kill the REST and WebUI threads).')
-            obj.console.print(Panel.fit(grid))
-            obj.console.log(f"")
-            obj.console.log(f"")
-            obj.console.log(f"")
-            obj.console.log(f"")
+            if web:
+                webui_thread = None
+                obj.console.log(f'Starting up Web UI on {host}:{webui_port}')
+                webui = WebServer(host, webui_port, host, rest_port)
+                webui_thread = threading.Thread(target=webui.run, name='NanoRC_WebUI')
+                webui_thread.start()
+                obj.console.log(f"")
+                obj.console.log(f"")
+                obj.console.log(f"")
+                obj.console.log(f"")
+                grid = Table(title='Web NanoRC', show_header=False, show_edge=False)
+                grid.add_column()
+                grid.add_row(f"Started Web UI, you can now connect to: [blue]{host}:{webui_port}[/blue],")
+                if 'np04' in host:
+                    grid.add_row(f"You probably need to set up a SOCKS proxy to lxplus:")
+                    grid.add_row("[blue]ssh -N -D 8080 your_cern_uname@lxtunnel.cern.ch[/blue] # on a different terminal window on your machine")
+                    grid.add_row(f'Make sure you set up browser SOCKS proxy with port 8080 too,')
+                    grid.add_row('on Chrome, \'Hotplate localhost SOCKS proxy setup\' works well).')
+                grid.add_row()
+                grid.add_row(f'[red]To stop this, ctrl-c [/red][bold red]twice[/bold red] (that will kill the REST and WebUI threads).')
+                obj.console.print(Panel.fit(grid))
+                obj.console.log(f"")
+                obj.console.log(f"")
+                obj.console.log(f"")
+                obj.console.log(f"")
+
     except Exception as e:
         logging.getLogger("cli").exception("Failed to build NanoRC")
         raise click.Abort()
@@ -143,9 +150,17 @@ def timingcli(ctx, obj, traceback, pm, loglevel, log_path, cfg_dumpdir, kerberos
     obj.rc = rc
     obj.shell = ctx.command
     rc.ls(False)
+
     if web:
         rest_thread.join()
         webui_thread.join()
+    
+    if tui:
+        from .tui import NanoRCTUI
+        tui = NanoRCTUI(host=host, rest_port=rest_port, timeout=rc.timeout, banner=Panel.fit(grid))
+        tui.run()
+        cleanup_rc()
+        ctx.exit(rc.return_code)
 
 
 from nanorc.common_commands import status, boot, conf, scrap, wait, terminate, start_shell, stop_run, shutdown
