+++ src/nanorc/cli.py
@@ -81,6 +81,7 @@ def updateLogLevel(loglevel):
 @click.option('--logbook-prefix', type=str, default="logbook", help='Prefix for the logbook file')
 @click.option('--pm', type=str, default="ssh://", help='Process manager, can be: ssh://, kind://, or k8s://np04-srv-015:31000, for example', callback=argval.validate_pm)
 @click.option('--web/--no-web', is_flag=True, default=False, help='whether to spawn webui')
+@click.option('--tui/--no-tui', is_flag=True, default=False, help='whether to use TUI')
 @accept_timeout(60)
 @click.option('--partition-number', type=int, default=0, help='Which partition number to run', callback=argval.validate_partition_number)
 @click.option('--web/--no-web', is_flag=True, default=False, help='whether to spawn WEBUI')
@@ -88,10 +89,12 @@ def updateLogLevel(loglevel):
 @click.argument('partition-label', type=str, callback=argval.validate_partition)
 @click.pass_obj
 @click.pass_context
-def cli(ctx, obj, traceback, loglevel, cfg_dumpdir, log_path, logbook_prefix, timeout, kerberos, partition_number, web, top_cfg, partition_label, pm):
+def cli(ctx, obj, traceback, loglevel, cfg_dumpdir, log_path, logbook_prefix, timeout, kerberos, partition_number, web, top_cfg, partition_label, tui, pm):
     obj.print_traceback = traceback
     credentials.user = 'user'
     ctx.command.shell.prompt = f'{credentials.user}@rc> '
+
+    obj.use_rest = web or tui
 
     grid = Table(title='Shonky NanoRC', show_header=False, show_edge=False)
     grid.add_column()
@@ -100,7 +103,8 @@ def cli(ctx, obj, traceback, loglevel, cfg_dumpdir, log_path, logbook_prefix, ti
     grid.add_row("  but trust it and it will betray you!")
     grid.add_row(f"Use it with care, {credentials.user}!")
 
-    obj.console.print(Panel.fit(grid))
+    if not tui:
+        obj.console.print(Panel.fit(grid))
 
     port_offset = 0 + partition_number * 500
     rest_port = 5005 + partition_number
@@ -111,7 +115,8 @@ def cli(ctx, obj, traceback, loglevel, cfg_dumpdir, log_path, logbook_prefix, ti
 
     rest_thread  = threading.Thread()
     webui_thread = threading.Thread()
-
+    if web and tui:
+        raise RuntimeError("cant have TUI and GUI at the same time")
     try:
         rc = NanoRC(
             console = obj.console,
@@ -133,7 +138,7 @@ def cli(ctx, obj, traceback, loglevel, cfg_dumpdir, log_path, logbook_prefix, ti
         add_common_cmds(ctx.command)
         add_custom_cmds(ctx.command, rc.execute_custom_command, rc.custom_cmd)
 
-        if web:
+        if web or tui:                  #The TUI also needs the rest thread
             host = socket.gethostname()
 
             rc_context.obj = obj
@@ -149,35 +154,36 @@ def cli(ctx, obj, traceback, loglevel, cfg_dumpdir, log_path, logbook_prefix, ti
             rest_thread.start()
             obj.console.log(f"Started RESTAPI")
 
-            webui_thread = None
-            obj.console.log(f'Starting up Web UI on {host}:{webui_port}')
-            webui = WebServer(host, webui_port, host, rest_port)
-            webui_thread = threading.Thread(target=webui.run, name='NanoRC_WebUI')
-            webui_thread.start()
-            obj.console.log(f"")
-            obj.console.log(f"")
-            obj.console.log(f"")
-            obj.console.log(f"")
-            grid = Table(title='Web NanoRC', show_header=False, show_edge=False)
-            grid.add_column()
-            grid.add_row(f"Started Web UI, you can now connect to: [blue]{host}:{webui_port}[/blue],")
-            if 'np04' in host:
-                grid.add_row(f"You probably need to set up a SOCKS proxy to lxplus:")
-                grid.add_row("[blue]ssh -N -D 8080 your_cern_uname@lxtunnel.cern.ch[/blue] # on a different terminal window on your machine")
-                grid.add_row(f'Make sure you set up browser SOCKS proxy with port 8080 too,')
-                grid.add_row('on Chrome, \'Hotplate localhost SOCKS proxy setup\' works well).')
-            elif 'lxplus' in host:
-                grid.add_row(f"You probably need to set up a SOCKS proxy to lxplus:")
-                grid.add_row(f"[blue]ssh -N -D 8080 your_cern_uname@{host}[/blue] # on a different terminal window on your machine")
-                grid.add_row(f'Make sure you set up browser SOCKS proxy with port 8080 too,')
-                grid.add_row('on Chrome, \'Hotplate localhost SOCKS proxy setup\' works well).')
-            grid.add_row()
-            grid.add_row(f'[red]To stop this, ctrl-c [/red][bold red]twice[/bold red] (that will kill the REST and WebUI threads).')
-            obj.console.print(Panel.fit(grid))
-            obj.console.log(f"")
-            obj.console.log(f"")
-            obj.console.log(f"")
-            obj.console.log(f"")
+            if web:
+                webui_thread = None
+                obj.console.log(f'Starting up Web UI on {host}:{webui_port}')
+                webui = WebServer(host, webui_port, host, rest_port)
+                webui_thread = threading.Thread(target=webui.run, name='NanoRC_WebUI')
+                webui_thread.start()
+                obj.console.log(f"")
+                obj.console.log(f"")
+                obj.console.log(f"")
+                obj.console.log(f"")
+                grid = Table(title='Web NanoRC', show_header=False, show_edge=False)
+                grid.add_column()
+                grid.add_row(f"Started Web UI, you can now connect to: [blue]{host}:{webui_port}[/blue],")
+                if 'np04' in host:
+                    grid.add_row(f"You probably need to set up a SOCKS proxy to lxplus:")
+                    grid.add_row("[blue]ssh -N -D 8080 your_cern_uname@lxtunnel.cern.ch[/blue] # on a different terminal window on your machine")
+                    grid.add_row(f'Make sure you set up browser SOCKS proxy with port 8080 too,')
+                    grid.add_row('on Chrome, \'Hotplate localhost SOCKS proxy setup\' works well).')
+                elif 'lxplus' in host:
+                    grid.add_row(f"You probably need to set up a SOCKS proxy to lxplus:")
+                    grid.add_row(f"[blue]ssh -N -D 8080 your_cern_uname@{host}[/blue] # on a different terminal window on your machine")
+                    grid.add_row(f'Make sure you set up browser SOCKS proxy with port 8080 too,')
+                    grid.add_row('on Chrome, \'Hotplate localhost SOCKS proxy setup\' works well).')
+                grid.add_row()
+                grid.add_row(f'[red]To stop this, ctrl-c [/red][bold red]twice[/bold red] (that will kill the REST and WebUI threads).')
+                obj.console.print(Panel.fit(grid))
+                obj.console.log(f"")
+                obj.console.log(f"")
+                obj.console.log(f"")
+                obj.console.log(f"")
 
 
     except Exception as e:
@@ -199,6 +205,13 @@ def cli(ctx, obj, traceback, loglevel, cfg_dumpdir, log_path, logbook_prefix, ti
     if web:
         rest_thread.join()
         webui_thread.join()
+
+    if tui:
+        from .tui import NanoRCTUI
+        tui = NanoRCTUI(host=host, rest_port=rest_port, timeout=rc.timeout, banner=Panel.fit(grid))
+        tui.run()
+        cleanup_rc()
+        ctx.exit(rc.return_code)
